BE/requirements.txt
anyio==3.5.0
asgiref==3.5.0
Authlib==0.15.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
dnspython==2.2.1
fastapi==0.75.0
h11==0.13.0
httptools==0.3.0
idna==3.3
motor==2.5.1
pycparser==2.21
pydantic==1.9.0
pymongo==3.12.3
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
pyuseragents~=1.0.5
boto3~=1.21.22
reportlab~=3.6.8

# Testing
pytest==7.1.1
pytest-asyncio==0.18.2

# Utils
PyJWT==2.3.0
loguru==0.6.0